requirements.txt
flask==2.0.1
pymongo==3.12.0
flask-pymongo==2.3.0
python-dotenv==0.19.0
flask-bcrypt==0.7.1